Make Gutterless List demo simpler
Changing name of demo to Gutterless List to follow the pattern set for most List demos.

Eliminated Card components to focus on List.
